[FEAT] MainLayout 컴포넌트 분리 및 Footer 정보 추가 | Split MainLayout component and added Footer information TALEARNT-491
