+++ src/components/common/icons/LogoIcon/LogoIcon.tsx
@@ -1,15 +1,12 @@
-import { classNames } from "@shared/utils/classNames";
-
 import { CommonIconProps } from "@components/common/icons/icon.type";
 
 /**
  * 로고 아이콘
  *
- * @param {string | undefined} className
- * @param {Omit<React.ComponentProps<"svg">, "className">} props
+ * @param {React.ComponentProps<"svg">} props
  * @constructor
  */
-function LogoIcon({ className, ...props }: CommonIconProps) {
+function LogoIcon(props: CommonIconProps) {
   return (
     <svg
       width="170"
@@ -17,18 +14,16 @@ function LogoIcon({ className, ...props }: CommonIconProps) {
       viewBox="0 0 170 35"
       fill="none"
       xmlns="http://www.w3.org/2000/svg"
-      className={classNames(className)}
       {...props}
     >
-      <rect width="170" height="35" fill="white" />
+      <rect width="170" height="35" />
       <path
         d="M46.3308 31.5437L44.7816 20.5246C44.2579 15.3242 39.8721 11.2656 34.5334 11.2656C28.8457 11.2656 24.2344 15.8769 24.2344 21.5647V24.1395C24.2344 29.8272 28.8457 34.4385 34.5334 34.4385C37.0718 34.4385 39.3993 33.5148 41.1958 31.9947L41.2322 32.2565C41.4286 33.6675 42.7305 34.6494 44.1415 34.4458C45.5526 34.2494 46.5345 32.9475 46.3308 31.5364V31.5437ZM39.683 21.572V24.1467C39.683 26.9906 37.3773 29.2963 34.5334 29.2963C31.6896 29.2963 29.3839 26.9906 29.3839 24.1467V21.572C29.3839 18.7281 31.6896 16.4224 34.5334 16.4224C37.1155 16.4224 39.2538 18.328 39.6175 20.8083C39.6175 20.8083 39.6175 20.8228 39.6175 20.8301L39.6539 21.0919C39.6684 21.2519 39.6757 21.4119 39.6757 21.572H39.683Z"
         fill="#1B76FF"
       />
       <path
         d="M39.6099 22.3016C39.7408 21.4142 39.959 20.556 40.2572 19.7341C40.2717 19.6832 40.2863 19.625 40.2863 19.5668C40.2863 19.254 40.0317 18.9922 39.7117 18.9922C39.6681 18.9922 39.6244 18.9922 39.5808 19.0067C39.4644 19.0358 39.3335 19.0504 39.2098 19.0504C39.0571 19.0504 38.9116 19.0286 38.7734 18.9922L38.9989 19.6759L39.4717 21.6688L39.6099 22.2943V22.3016Z"
-        fill="white"
-        className={""}
+        fill="#FFFFFF"
       />
       <path
         d="M105.089 31.5437L103.539 20.5246C103.016 15.3242 98.6299 11.2656 93.2912 11.2656C87.6035 11.2656 82.9922 15.8769 82.9922 21.5647V24.1395C82.9922 29.8272 87.6035 34.4385 93.2912 34.4385C95.8296 34.4385 98.1571 33.5148 99.9536 31.9947L99.99 32.2565C100.186 33.6675 101.488 34.6494 102.899 34.4458C104.31 34.2494 105.292 32.9475 105.089 31.5364V31.5437ZM98.4481 21.572V24.1467C98.4481 26.9906 96.1424 29.2963 93.2985 29.2963C90.4546 29.2963 88.149 26.9906 88.149 24.1467V21.572C88.149 18.7281 90.4546 16.4224 93.2985 16.4224C95.8806 16.4224 98.0189 18.328 98.3826 20.8083C98.3826 20.8083 98.3826 20.8228 98.3826 20.8301L98.419 21.0919C98.4335 21.2519 98.4408 21.4119 98.4408 21.572H98.4481Z"
@@ -60,7 +55,7 @@ function LogoIcon({ className, ...props }: CommonIconProps) {
       />
       <path
         d="M39.6595 21.9657C39.6595 21.9657 39.6595 21.9439 39.6668 21.9294V21.573C39.6668 20.6493 39.4195 19.7837 38.9976 19.0346C38.9758 18.9982 38.954 18.9618 38.9322 18.9255C38.9322 18.9255 38.9249 18.9109 38.9176 18.8964C38.5321 18.2636 38.0157 17.7254 37.4048 17.3181C37.3684 17.2962 37.332 17.2672 37.2957 17.2453C37.623 16.5107 37.8557 15.6597 37.9357 14.7433C37.9866 14.2123 37.9866 13.6887 37.9357 13.1941C37.9066 12.9322 37.8121 12.6922 37.6521 12.4958C37.3175 12.0667 36.7065 12.0085 36.2992 12.3722C36.1101 12.5395 35.9647 12.7577 35.8919 13.0122C35.7537 13.4923 35.6592 14.0014 35.6083 14.5324C35.5428 15.2452 35.5646 15.9361 35.6737 16.5762C35.6664 16.5762 35.6592 16.5762 35.6446 16.5762C35.5646 16.5616 35.4919 16.5471 35.4119 16.5325C35.4046 16.5325 35.3973 16.5325 35.3901 16.5325C35.1064 16.4816 34.8227 16.4598 34.5245 16.4598C34.3354 16.4598 34.1463 16.4744 33.9645 16.4889C33.8336 15.9943 33.6372 15.4852 33.3753 14.9906C33.1499 14.5615 32.8953 14.1687 32.6189 13.8123C32.4734 13.6232 32.2771 13.4923 32.0589 13.4195C31.5934 13.2595 31.0988 13.5141 30.9679 13.9869C30.9097 14.2051 30.9024 14.4451 30.9679 14.6706C31.0988 15.0997 31.2733 15.5361 31.4988 15.9652C31.7025 16.3507 31.9279 16.7071 32.168 17.0271C30.5096 17.8781 29.375 19.6092 29.375 21.5948V24.1623C29.375 26.9989 31.6734 29.3045 34.5173 29.3045C37.3611 29.3045 39.6595 27.0062 39.6595 24.1623V21.9803V21.9657Z"
-        fill="white"
+        fill="#FFFFFF"
       />
       <path
         d="M36.7929 16.9584C36.5893 16.8566 36.3711 16.7693 36.1529 16.6966C36.0292 16.2529 35.9856 15.7438 36.0365 15.2056C36.0656 14.9146 36.1165 14.6455 36.1892 14.3837C36.2256 14.2455 36.3056 14.1291 36.4074 14.0418C36.6256 13.8454 36.9602 13.8818 37.1348 14.1073C37.2148 14.2164 37.2729 14.3473 37.2875 14.4855C37.3093 14.7546 37.3166 15.031 37.2875 15.3219C37.2293 15.9402 37.0547 16.5075 36.7929 16.9584Z"
@@ -84,12 +79,11 @@ function LogoIcon({ className, ...props }: CommonIconProps) {
       />
       <path
         d="M88.2155 22.3016C88.0846 21.4142 87.8664 20.556 87.5682 19.7341C87.5536 19.6832 87.5391 19.625 87.5391 19.5668C87.5391 19.254 87.7936 18.9922 88.1137 18.9922C88.1573 18.9922 88.2009 18.9922 88.2446 19.0067C88.361 19.0358 88.4919 19.0504 88.6155 19.0504C88.7683 19.0504 88.9137 19.0286 89.0519 18.9922L88.8264 19.6759L88.3537 21.6688L88.2155 22.2943V22.3016Z"
-        fill="white"
-        className={""}
+        fill="#FFFFFF"
       />
       <path
         d="M88.1635 21.9657C88.1635 21.9657 88.1635 21.9439 88.1562 21.9294V21.573C88.1562 20.6493 88.4035 19.7837 88.8254 19.0346C88.8472 18.9982 88.869 18.9618 88.8909 18.9255C88.8909 18.9255 88.8981 18.9109 88.9054 18.8964C89.2909 18.2636 89.8073 17.7254 90.4183 17.3181C90.4546 17.2962 90.491 17.2672 90.5274 17.2453C90.2001 16.5107 89.9673 15.6597 89.8873 14.7433C89.8364 14.2123 89.8364 13.6887 89.8873 13.1941C89.9164 12.9322 90.011 12.6922 90.171 12.4958C90.5055 12.0667 91.1165 12.0085 91.5238 12.3722C91.7129 12.5395 91.8584 12.7577 91.9311 13.0122C92.0693 13.4923 92.1639 14.0014 92.2148 14.5324C92.2802 15.2452 92.2584 15.9361 92.1493 16.5762C92.1566 16.5762 92.1639 16.5762 92.1784 16.5762C92.2584 16.5616 92.3312 16.5471 92.4112 16.5325C92.4184 16.5325 92.4257 16.5325 92.433 16.5325C92.7166 16.4816 93.0003 16.4598 93.2985 16.4598C93.4876 16.4598 93.6767 16.4744 93.8586 16.4889C93.9895 15.9943 94.1859 15.4852 94.4477 14.9906C94.6732 14.5615 94.9277 14.1687 95.2041 13.8123C95.3496 13.6232 95.546 13.4923 95.7642 13.4195C96.2297 13.2595 96.7243 13.5141 96.8552 13.9869C96.9134 14.2051 96.9206 14.4451 96.8552 14.6706C96.7243 15.0997 96.5497 15.5361 96.3242 15.9652C96.1206 16.3507 95.8951 16.7071 95.6551 17.0271C97.3134 17.8781 98.448 19.6092 98.448 21.5948V24.1623C98.448 26.9989 96.1497 29.3045 93.3058 29.3045C90.4619 29.3045 88.1635 27.0062 88.1635 24.1623V22.1185V21.9803V21.9657Z"
-        fill="white"
+        fill="#FFFFFF"
       />
       <path
         d="M91.0372 16.9584C91.2409 16.8566 91.4591 16.7693 91.6773 16.6966C91.8009 16.2529 91.8446 15.7438 91.7936 15.2056C91.7645 14.9146 91.7136 14.6455 91.6409 14.3837C91.6045 14.2455 91.5245 14.1291 91.4227 14.0418C91.2045 13.8454 90.8699 13.8818 90.6954 14.1073C90.6154 14.2164 90.5572 14.3473 90.5426 14.4855C90.5208 14.7546 90.5135 15.031 90.5426 15.3219C90.6008 15.9402 90.7754 16.5075 91.0372 16.9584Z"
